Backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
